requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
pydantic-settings>=2.3,<3.0
aiofiles>=24.1,<25.0
filetype>=1.2,<2.0
httpx>=0.27,<1.0
faster-whisper>=1.1,<2.0
concurrent-log-handler>=0.9,<1.0
python-multipart>=0.0.9,<1.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
scenedetect[opencv]>=0.6.7,<1.0
pymysql>=1.1,<2.0
SQLAlchemy>=2.0,<3.0
PyJWT>=2.9,<3.0
alembic>=1.13,<2.0
